+++ contentbank/classes/external/delete_content.php
@@ -14,21 +14,32 @@
 // You should have received a copy of the GNU General Public License
 // along with Moodle.  If not, see <http://www.gnu.org/licenses/>.
 
-namespace core_contentbank\external;
-
-use core_contentbank\contentbank;
-use core_external\external_api;
-use core_external\external_function_parameters;
-use core_external\external_multiple_structure;
-use core_external\external_single_structure;
-use core_external\external_value;
-use core_external\external_warnings;
-
 /**
  * This is the external method for deleting a content.
  *
  * @package    core_contentbank
  * @since      Moodle 3.9
+ * @copyright  2020 Sara Arjona <sara@moodle.com>
+ * @license    http://www.gnu.org/copyleft/gpl.html GNU GPL v3 or later
+ */
+
+namespace core_contentbank\external;
+
+defined('MOODLE_INTERNAL') || die();
+
+global $CFG;
+require_once($CFG->libdir . '/externallib.php');
+
+use external_api;
+use external_function_parameters;
+use external_multiple_structure;
+use external_single_structure;
+use external_value;
+use external_warnings;
+
+/**
+ * This is the external method for deleting a content.
+ *
  * @copyright  2020 Sara Arjona <sara@moodle.com>
  * @license    http://www.gnu.org/copyleft/gpl.html GNU GPL v3 or later
  */
@@ -59,31 +70,34 @@ class delete_content extends external_api {
         $warnings = [];
 
         $params = self::validate_parameters(self::execute_parameters(), ['contentids' => $contentids]);
-        $cb = new contentbank();
         foreach ($params['contentids'] as $contentid) {
             try {
                 $record = $DB->get_record('contentbank_content', ['id' => $contentid], '*', MUST_EXIST);
-                $content = $cb->get_content_from_id($record->id);
-                $contenttype = $content->get_content_type_instance();
-                $context = \context::instance_by_id($record->contextid, MUST_EXIST);
-                self::validate_context($context);
-                // Check capability.
-                if ($contenttype->can_delete($content)) {
-                    // This content can be deleted.
-                    if (!$contenttype->delete_content($content)) {
+                $contenttypeclass = "\\$record->contenttype\\contenttype";
+                if (class_exists($contenttypeclass)) {
+                    $context = \context::instance_by_id($record->contextid, MUST_EXIST);
+                    self::validate_context($context);
+                    $contenttype = new $contenttypeclass($context);
+                    $contentclass = "\\$record->contenttype\\content";
+                    $content = new $contentclass($record);
+                    // Check capability.
+                    if ($contenttype->can_delete($content)) {
+                        // This content can be deleted.
+                        if (!$contenttype->delete_content($content)) {
+                            $warnings[] = [
+                                'item' => $contentid,
+                                'warningcode' => 'contentnotdeleted',
+                                'message' => get_string('contentnotdeleted', 'core_contentbank')
+                            ];
+                        }
+                    } else {
+                        // The user has no permission to delete this content.
                         $warnings[] = [
                             'item' => $contentid,
-                            'warningcode' => 'contentnotdeleted',
-                            'message' => get_string('contentnotdeleted', 'core_contentbank')
+                            'warningcode' => 'nopermissiontodelete',
+                            'message' => get_string('nopermissiontodelete', 'core_contentbank')
                         ];
                     }
-                } else {
-                    // The user has no permission to delete this content.
-                    $warnings[] = [
-                        'item' => $contentid,
-                        'warningcode' => 'nopermissiontodelete',
-                        'message' => get_string('nopermissiontodelete', 'core_contentbank')
-                    ];
                 }
             } catch (\moodle_exception $e) {
                 // The content or the context don't exist.
